Int to Long, typo fixed

+++ train.py
@@ -41,8 +41,8 @@ for epoch in range(args.n_epochs):
     for i, (x, y, n) in enumerate(tr_iterator()):
         xin = Variable(torch.FloatTensor(x))
         xout = Variable(torch.FloatTensor(x))
-        y = Variable(torch.IntTensor(y))
-        n = Variable(torch.IntTensor(n))
+        y = Variable(torch.LongTensor(y))
+        n = Variable(torch.LongTensor(n))
 
         mu2, qz2_x, z2, qz1_x, z1, px_z, x_sample = fhvae(xin, xout, y)
